Rename Gainput to SDLInput and update project wizzard in launcher (#32)
