--- GloryEngine/Engine/Modules/GloryGainput/GainputModule.cpp
@@ -1,26 +1,0 @@
-#include "GainputModule.h"
-#include <Debug.h>
-#include <Engine.h>
-
-namespace Glory
-{
-	GainputModule::GainputModule()
-	{
-	}
-
-	GainputModule::~GainputModule()
-	{
-	}
-
-	void GainputModule::OnInitialize()
-	{
-	}
-
-	void GainputModule::OnCleanup()
-	{
-	}
-
-	void GainputModule::OnUpdate()
-	{
-	}
-}
