--- GloryEngine/Engine/Modules/GloryGainput/Module.yaml
@@ -1,7 +1,0 @@
-Name: "Gainput"
-Type: Input
-#Scripting:
-#- Language: "csharp"
-#  Extension: "GloryGainputMonoAPI"
-#Editor:
-# Extensions: [GainputEditorExtension]
