+++ GloryEngine/Engine/Modules/GlorySDLInput/GlorySDLInput.cpp
@@ -1,9 +1,9 @@
 #include <GloryContext.h>
-#include "GainputModule.h"
-#include "GloryGainput.h"
+#include "SDLInputModule.h"
+#include "GlorySDLInput.h"
 
 GLORY_API Glory::Module* OnLoadModule(Glory::GloryContext* pContext)
 {
     Glory::GloryContext::SetContext(pContext);
-    return new Glory::GainputModule();
+    return new Glory::SDLInputModule();
 }
