+++ GloryEngine/Engine/Modules/GlorySDLInput/Module.yaml
@@ -1,0 +1,2 @@
+Name: "SDL Input"
+Type: Input
