+++ GloryEngine/Engine/Modules/GlorySDLInput/SDLInputModule.cpp
@@ -1,0 +1,26 @@
+#include "SDLInputModule.h"
+#include <Debug.h>
+#include <Engine.h>
+
+namespace Glory
+{
+	SDLInputModule::SDLInputModule()
+	{
+	}
+
+	SDLInputModule::~SDLInputModule()
+	{
+	}
+
+	void SDLInputModule::OnInitialize()
+	{
+	}
+
+	void SDLInputModule::OnCleanup()
+	{
+	}
+
+	void SDLInputModule::OnUpdate()
+	{
+	}
+}
