+++ GloryEngine/premake5.lua
@@ -19,7 +19,7 @@ workspace "GloryEngine"
 	startproject "GloryEditorSDL"
 	startproject "GloryEditorOGL"
 	startproject "GloryMonoScripting"
-	startproject "GloryGainput"
+	startproject "GlorySDLInput"
 	--startproject "GloryEditorVulkan"
 	--startproject "GloryVulkanGraphics"
 	
@@ -140,7 +140,7 @@ group "Engine/Modules"
 	include "Engine/Modules/GlorySDLImage"
 	include "Engine/Modules/GlorySDLWindow"
 	include "Engine/Modules/GloryMonoScripting"
-	include "Engine/Modules/GloryGainput"
+	include "Engine/Modules/GlorySDLInput"
 	--include "GloryVulkanGraphics"
 group ""
